requirements.txt
python-dotenv>=1.0.1
langchain>=0.3.19
langchain-community>=0.3.18
pyodbc>=5.2.0
1.65.2>=1.65.2
langchain-openai>=0.3.7
langchain-core>=0.3.40
langchain-openai>=0.3.7
langchain-text-splitters>=0.3.6
langchainhub>=0.1.21